+++ CHANGELOG.md
@@ -9,8 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added an Equality comparison overload to ByteArray
 
+### Fixed
+- Added template restrictions for iterator constructor to avoid ambiguity between constructors
+
 ### Changed
 - Updated Readme.md with examples
+
 
 
 ## [0.1.0-alpha.1]